+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - [Feature: support spring-retry router](https://github.com/Tencent/spring-cloud-tencent/pull/633)
 - [Bugfix: fix throw npe when router context is null](https://github.com/Tencent/spring-cloud-tencent/pull/634)
 - [feat:Transfer http headers specified by environment variables](https://github.com/Tencent/spring-cloud-tencent/pull/637)
+- [feat:support read config file from local file system.](https://github.com/Tencent/spring-cloud-tencent/pull/643)